Added weather update date. Moved TableViewController functions
