+++ Model.xcdatamodeld/Model v2.xcdatamodel/contents
@@ -13,8 +13,9 @@
         <attribute name="urlWeatherCityNumberDate" optional="YES" attributeType="String"/>
         <attribute name="weather_state_abbr" optional="YES" attributeType="String"/>
         <attribute name="weather_state_name" optional="YES" attributeType="String"/>
+        <attribute name="weatherUpdateTimestamp" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
     </entity>
     <elements>
-        <element name="Item" positionX="-9" positionY="36" width="128" height="223"/>
+        <element name="Item" positionX="-9" positionY="36" width="128" height="238"/>
     </elements>
 </model>
