+++ TableViewController.swift
@@ -10,19 +10,6 @@ import UIKit
 import CoreData
 import SDWebImage
 
-protocol ReloadProtocol: AnyObject {
-    func reload()
-}
-
-extension TableViewController: ReloadProtocol {
-    func reload() {
-        fetchData()
-        DispatchQueue.main.async {
-            self.tableView.reloadData()
-        }
-    }
-}
-
 class TableViewController: UITableViewController {
     
     let cellId = "Cell"
@@ -52,14 +39,7 @@ class TableViewController: UITableViewController {
         fetchData()
     }
     
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        super.prepare(for: segue, sender: sender)
-        if let id = segue.identifier, id == "addItem", let vc = segue.destination as? AddItemViewController {
-            vc.delegate = self
-        }
-    }
-
-    // MARK: - tableview data source
+    // MARK: - Tableview data source
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return items.count
     }
@@ -75,7 +55,6 @@ class TableViewController: UITableViewController {
         cell.timeLabel.text      = timeStamp
         let wStateAbbr = item.weather_state_abbr  ?? "City unknown"
         let stateImageURL = "https://www.metaweather.com/static/img/weather/png/64/\(wStateAbbr).png"
-        print("\(stateImageURL) with \(item.weather_state_abbr) # \(item.cityNumber) abbr \(wStateAbbr)")
         cell.weatherImage.sd_setImage(with: URL(string: stateImageURL), placeholderImage: UIImage(named: "placeholderImage"))
         cell.weatherState.text   = item.weather_state_name
         cell.occurrenceDate.text = item.occurrenceDate
